requirements.txt
CherryPy
bcrypt
requests
mysql-connector-python
